Moved loadHTML to utilities
