+++ server.js
@@ -1,7 +1,11 @@
 const fs = require('fs')
 const express = require('express')
 const morgan = require('morgan')
-const databases = require('./database.js')
+const databases = require('./app/database.js')
+const utilities = require('./app/utilities.js')
+
+
+const loadHTML = utilities.loadHtml
 
 const args = require('minimist')(process.argv)
 
@@ -25,23 +29,23 @@ function loadFileAsText(filename) {
   return fs.readFileSync(filename, {encoding:"utf-8", flags:'r'})
 }
 
-/**
- * Load the html file from a template and injects body into a placeholder div
- * Default directory for lookup is "./www"
- * The placeholder div will be replaced
- * Do not include ".html" in filenames
- * 
- * @param {String} template the filename of the template html file (do not include .html in name)
- * @param {String} body     the filename of the body html file (do not include .html in name)
- * @param {*} placeholder   the id of the placeholder div used to replace content
- * @returns 
- */
-function loadHTML (template, body, placeholder, directory){
-  let path = directory || "./www"
-  let html = loadFileAsText(`${path}/${template}.html`)
-  let content = loadFileAsText(`${path}/${body}.html`)
-  return html.replace(`<div id="${placeholder}"></div>`, content)
-}
+// /**
+//  * Load the html file from a template and injects body into a placeholder div
+//  * Default directory for lookup is "./www"
+//  * The placeholder div will be replaced
+//  * Do not include ".html" in filenames
+//  * 
+//  * @param {String} template the filename of the template html file (do not include .html in name)
+//  * @param {String} body     the filename of the body html file (do not include .html in name)
+//  * @param {*} placeholder   the id of the placeholder div used to replace content
+//  * @returns 
+//  */
+// function loadHTML (template, body, placeholder, directory){
+//   let path = directory || "./www"
+//   let html = loadFileAsText(`${path}/${template}.html`)
+//   let content = loadFileAsText(`${path}/${body}.html`)
+//   return html.replace(`<div id="${placeholder}"></div>`, content)
+// }
 
 
 // Endpoint for the main page - this is a test page right now
